requirements.txt
Django==5.2.12
PyJWT==2.12.1
asgiref==3.11.1
celery==5.6.3
django-cors-headers==4.9.0
django-environ==0.13.0
django-redis==6.0.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
numpy==2.2.6
pgvector==0.4.2
psycopg2-binary==2.9.11
redis==7.4.0
sqlparse==0.5.5
typing_extensions==4.15.0
